# backend/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "AcademiaConnect API",
                 "description":  "Backend API for AcademiaConnect SIH 2026",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/dashboard/student/{user_id}":  {
                                                       "get":  {
                                                                   "tags":  [
                                                                                "Dashboard"
                                                                            ],
                                                                   "summary":  "Student Dashboard",
                                                                   "operationId":  "student_dashboard_dashboard_student__user_id__get",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "name":  "user_id",
                                                                                          "in":  "path",
                                                                                          "required":  true,
                                                                                          "schema":  {
                                                                                                         "type":  "integer",
                                                                                                         "title":  "User Id"
                                                                                                     }
                                                                                      }
                                                                                  ],
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  "Successful Response",
                                                                                                 "content":  {
                                                                                                                 "application/json":  {
                                                                                                                                          "schema":  {

                                                                                                                                                     }
                                                                                                                                      }
                                                                                                             }
                                                                                             },
                                                                                     "422":  {
                                                                                                 "description":  "Validation Error",
                                                                                                 "content":  {
                                                                                                                 "application/json":  {
                                                                                                                                          "schema":  {
                                                                                                                                                         "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                     }
                                                                                                                                      }
                                                                                                             }
                                                                                             }
                                                                                 }
                                                               }
                                                   },
                  "/dashboard/company/{user_id}":  {
                                                       "get":  {
                                                                   "tags":  [
                                                                                "Dashboard"
                                                                            ],
                                                                   "summary":  "Company Dashboard",
                                                                   "operationId":  "company_dashboard_dashboard_company__user_id__get",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "name":  "user_id",
                                                                                          "in":  "path",
                                                                                          "required":  true,
                                                                                          "schema":  {
                                                                                                         "type":  "integer",
                                                                                                         "title":  "User Id"
                                                                                                     }
                                                                                      }
                                                                                  ],
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  "Successful Response",
                                                                                                 "content":  {
                                                                                                                 "application/json":  {
                                                                                                                                          "schema":  {

                                                                                                                                                     }
                                                                                                                                      }
                                                                                                             }
                                                                                             },
                                                                                     "422":  {
                                                                                                 "description":  "Validation Error",
                                                                                                 "content":  {
                                                                                                                 "application/json":  {
                                                                                                                                          "schema":  {
                                                                                                                                                         "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                     }
                                                                                                                                      }
                                                                                                             }
                                                                                             }
                                                                                 }
                                                               }
                                                   },
                  "/dashboard/institution/{user_id}":  {
                                                           "get":  {
                                                                       "tags":  [
                                                                                    "Dashboard"
                                                                                ],
                                                                       "summary":  "Institution Dashboard",
                                                                       "operationId":  "institution_dashboard_dashboard_institution__user_id__get",
                                                                       "parameters":  [
                                                                                          {
                                                                                              "name":  "user_id",
                                                                                              "in":  "path",
                                                                                              "required":  true,
                                                                                              "schema":  {
                                                                                                             "type":  "integer",
                                                                                                             "title":  "User Id"
                                                                                                         }
                                                                                          }
                                                                                      ],
                                                                       "responses":  {
                                                                                         "200":  {
                                                                                                     "description":  "Successful Response",
                                                                                                     "content":  {
                                                                                                                     "application/json":  {
                                                                                                                                              "schema":  {

                                                                                                                                                         }
                                                                                                                                          }
                                                                                                                 }
                                                                                                 },
                                                                                         "422":  {
                                                                                                     "description":  "Validation Error",
                                                                                                     "content":  {
                                                                                                                     "application/json":  {
                                                                                                                                              "schema":  {
                                                                                                                                                             "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                         }
                                                                                                                                          }
                                                                                                                 }
                                                                                                 }
                                                                                     }
                                                                   }
                                                       },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
...
                                       "ValidationError":  {
                                                               "properties":  {
                                                                                  "loc":  {
                                                                                              "items":  {
                                                                                                            "anyOf":  [
                                                                                                                          {
                                                                                                                              "type":  "string"
                                                                                                                          },
                                                                                                                          {
                                                                                                                              "type":  "integer"
                                                                                                                          }
                                                                                                                      ]
                                                                                                        },
                                                                                              "type":  "array",
                                                                                              "title":  "Location"
                                                                                          },
                                                                                  "msg":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Message"
                                                                                          },
                                                                                  "type":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Error Type"
                                                                                           }
                                                                              },
                                                               "type":  "object",
                                                               "required":  [
                                                                                "loc",
                                                                                "msg",
                                                                                "type"
                                                                            ],
                                                               "title":  "ValidationError"
                                                           },
...
                                   }
                   }
}